+++ tests/test_supremo_lite.py
@@ -97,7 +97,7 @@ class TestSupremoLite(unittest.TestCase):
         self.assertEqual(len(metadata), expected_alt_count)
         self.assertEqual(
             list(metadata.columns),
-            ["chrom", "window_start", "window_end", "variant_pos0", "ref", "alt"],
+            ["chrom", "window_start", "window_end", "variant_offset0", "ref", "alt"],
         )
 
         # Verify some mutations